Move program check to common area
