+++ server/service/src/invoice/inventory_adjustment/adjust_existing_stock/validate.rs
@@ -1,6 +1,6 @@
 use repository::{RepositoryError, StockLine, StorageConnection};
 
-use crate::common_stock::{check_stock_line_exists, CommonStockLineError};
+use crate::common::{check_stock_line_exists, CommonStockLineError};
 use crate::invoice::inventory_adjustment::adjust_existing_stock::AdjustmentType;
 
 use crate::stocktake_line::validate::{check_active_adjustment_reasons, check_reason_is_valid};
