+++ server/service/src/vaccine_course/validate.rs
@@ -3,8 +3,7 @@ use repository::{
         vaccine_course::{VaccineCourseFilter, VaccineCourseRepository},
         vaccine_course_row::{VaccineCourseRow, VaccineCourseRowRepository},
     },
-    EqualFilter, ProgramRow, ProgramRowRepository, RepositoryError, StorageConnection,
-    StringFilter,
+    EqualFilter, RepositoryError, StorageConnection, StringFilter,
 };
 
 use super::update::VaccineCourseDoseInput;
@@ -14,13 +13,6 @@ pub fn check_vaccine_course_exists(
     connection: &StorageConnection,
 ) -> Result<Option<VaccineCourseRow>, RepositoryError> {
     VaccineCourseRowRepository::new(connection).find_one_by_id(id)
-}
-
-pub fn check_program_exists(
-    id: &str,
-    connection: &StorageConnection,
-) -> Result<Option<ProgramRow>, RepositoryError> {
-    ProgramRowRepository::new(connection).find_one_by_id(id)
 }
 
 pub fn check_vaccine_course_name_exists_for_program(
